bugfix: avoid string-set! error when peek if #!eof
depending on when peek is called (which varies on different versions
of CHICKEN, it seems), it's possible to end up with a peek variable
being #!eof.

this crashes when we read the string-set! in read-string! as #!eof is,
of course, not a valid character.

this seems to fix the problem.